app/requirements1.txt
amqp==5.2.0
asgiref==3.8.1
av==12.3.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
crispy-bootstrap5==2024.10
ctranslate2==4.4.0
Django==5.1.2
django-allauth==65.0.2
django-crispy-forms==2.3
faster-whisper==1.0.3
filelock==3.16.1
flatbuffers==24.3.25
flower==2.0.1
fsspec==2024.9.0
gunicorn==23.0.0
huggingface-hub==0.25.2
humanfriendly==10.0
humanize==4.11.0
idna==3.10
Jinja2==3.1.4
kombu==5.4.2
llvmlite==0.43.0
lxml==5.3.0
MarkupSafe==3.0.1
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.19.2
openai-whisper==20240930
packaging==24.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==5.28.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
regex==2024.9.11
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
sympy==1.13.3
tiktoken==0.8.0
tokenizers==0.20.1
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
